requirements.txt
# Core
feedparser==6.0.11
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
python-dateutil==2.9.0.post0
pytz==2024.1
APScheduler==3.10.4
pydantic==2.8.2
PyYAML==6.0.2
httpx==0.27.2
beautifulsoup4==4.12.3
tqdm==4.66.5

fastapi==0.112.2
uvicorn==0.30.6
